parser/requirements.txt
requests==2.31.0
python-dateutil==2.8.2
black==24.2.0
pytz==2024.1
pika==1.3.2
